Added scroll for robots and add and remove buttons
