+++ robofriends/src/components/RoboCard.js
