+++ robofriends/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import * as serviceWorker from './serviceWorker';
-import App from "./App";
+import App from "./containers/App";
 import "tachyons";
 
 ReactDOM.render(<App />, document.getElementById('root'));
